api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pyyaml
pytest
